Move rewriter utils to nlx module
